lib: let nl_object_identical() declare the same object as identical
nl_object_identical() only considers two objects as identical, if all
their ID attributes are set. That means, two objects that are
incomplete (with unset ID attributes) never compare identical.

That cannot be right. At least not if both objects are the very same
pointer.

That is also important for nl_cache_clear() and nl_cache_remove(). An
object must be identical to itself, otherwise it cannot be removed from
the cache anymore.